Hidden discount functionality

+++ Restaurant-Manager/Presentation/RemoveReservationPresentation.cs
@@ -16,6 +16,15 @@ public static class RemoveReservationPresentation {
         }
 
         long reservationID = ReservationLogic.ParseIDFromString(reservationChoice);
+        
+        if (reservationID == HiddenDiscount.HiddenCodeID)
+        {
+            Console.WriteLine("This reservation exists only to contain the hidden discount. You cannot remove this.");
+            AnsiConsole.Markup("[gray]Press any key to continue[/]");
+            Console.ReadKey();
+            return;
+        }
+        
         Database.DeleteReservationsTable(reservationID);
         AnsiConsole.WriteLine("Reservation Removed");
         AnsiConsole.Markup("[gray]Press any key to continue[/]");
